refactor(engine): created new building/shipBuilding to house related functions. migrated them over from building.ts
