+++ src/state/gameStore.ts
@@ -13,7 +13,8 @@ import { processAiTurn } from '../engine/ai';
 import { processCharacterCycles } from '../engine/character';
 import { processDiplomacy, enginePlayerDiploResponse, sendDiploRequest} from '../engine/diplomacy';
 import { generateGalaxy, generateStartingOrgs, generateStartingSpecies } from '../engine/galaxyGeneration';
-import { engineBuildShip, engineBuildBuilding, engineBuildMilShip, engineBuildPlanetoid, engineBuildAnchor } from '../engine/building';
+import { engineBuildBuilding, engineBuildPlanetoid, engineBuildAnchor } from '../engine/building';
+import { engineBuildShip, engineBuildMilShip } from '../engine/building/shipBuilding';
 import { colonizePlanetoid, beginPlanetoidSurvey, getHabitablesInSystem } from '../engine/colonization';
 import { engineAssignCharacter } from '../engine/character';
 import { shiftLanes } from '../engine/ecology';
